fx-deal-kilany/src/main/java/com/bloomberg/fxdealanalyzer/validator/CurrencyCodeValidator.java
package com.bloomberg.fxdealanalyzer.validator;

import com.bloomberg.fxdealanalyzer.annotation.ValidCurrency;

import javax.validation.ConstraintValidator;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import java.util.Currency;

public class CurrencyCodeValidator implements ConstraintValidator<ValidCurrency, String> {

    @Override
    public void initialize(ValidCurrency constraintAnnotation) {
    }

    @Override
    public boolean isValid(String currencyCode, ConstraintValidatorContext context) {
        if (currencyCode == null || currencyCode.isEmpty()) {
            return true;
        }
        try {
            Currency.getInstance(currencyCode);
            return true;
        } catch (IllegalArgumentException e) {
            return false;
        }
    }
}
